+++ Core/Arpschuino/libraries/arpschuino/examples/arpschuino/RF_Receiver/RF_Receiver.ino
@@ -32,7 +32,7 @@ SOFTPWM_DEFINE_OBJECT( 16 );
 #define DATA_PER_NODE (6) 
 #define NETWORKID (212)  //l'id du reseau commun pour toute les cartes
 #define NODEID (1) //l'id de la carte qui sert de gateway
-int band = RF12_868MHZ; //la frequence de l'emeteur
+uint16_t band = RF12_868MHZ; //la frequence de l'emeteur
 //RF12_915MHZ America  // RF12_433MHZ
 
 float frequency_setting = 868.00; //beetween 863,00 > 870,00
