+++ Core/Arpschuino/libraries/arpschuino/examples/arpschuino/emetteur_arpschuino_officiel/emetteur_arpschuino_officiel.ino
@@ -42,7 +42,7 @@ int address;
 #include <ArpRFLib.h>
 byte NODEID = 1;  //Adresse RF unique pour chaque machine de 1 à 30
 #define NETWORKID  212  //adresse du reseau commune à toute les machine
-uint8_t band = RF12_868MHZ; //frequence de l'emeteur RF12
+uint16_t band = RF12_868MHZ; //frequence de l'emeteur RF12
 //RF12_915MHZ America  // RF12_433MHZ
 
 float frequency_setting = 868.00; //beetween 863,00 > 870,00
