+++ Core/Arpschuino/libraries/arpschuino/examples/arpsensorsRF/MultiSenders/MultiSenders.ino
@@ -33,7 +33,7 @@ void read_sensor()
 #define ID 2//  unique pour chaque carte 2 droite, 3 pour gauche,etc...
 #define NETWORKID (212)  //l'id du reseau commun pour toute les cartes
 #define NODEID (1) //l'id de la carte qui sert de gateway
-int band = RF12_868MHZ; //la frequence de l'emeteur
+uint16_t band = RF12_868MHZ; //la frequence de l'emeteur
 //RF12_915MHZ America  // RF12_433MHZ
 
 float frequency_setting = 868.00; //beetween 863,00 > 870,00
